# swaggerSlot/go-server-server/go/slot.go
/*
 * Basic swagger example application
 *
 * A slot crud apis
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Slot struct {

	Id int32 `json:"id,omitempty"`

	SlotName string `json:"slot_name"`
}